@@ -13,7 +13,10 @@
 // limitations under the License.
 
 using System.Net;
+using RestSharp.Authenticators;
 using RestSharp.Extensions;
+
+// ReSharper disable ReplaceSubstringWithRangeIndexer
 // ReSharper disable UnusedAutoPropertyAccessor.Global
 
 namespace RestSharp;
@@ -22,8 +25,8 @@ namespace RestSharp;
 /// Container for data used to make requests
 /// </summary>
 public class RestRequest {
-    readonly Func<HttpResponseMessage, RestResponse>? _advancedResponseHandler;
-    readonly Func<Stream, Stream?>?                   _responseWriter;
+    readonly Func<HttpResponseMessage, RestRequest, RestResponse>? _advancedResponseHandler;
+    readonly Func<Stream, Stream?>?                                _responseWriter;
 
     /// <summary>
     /// Default constructor
@@ -47,19 +50,18 @@ public class RestRequest {
             var queryParams = ParseQuery(Resource.Substring(queryStringStart + 1));
             Resource = Resource.Substring(0, queryStringStart);
 
-            foreach (var param in queryParams)
-                this.AddQueryParameter(param.Key, param.Value, false);
+            foreach (var param in queryParams) this.AddQueryParameter(param.Key, param.Value, false);
         }
 
-        static IEnumerable<KeyValuePair<string, string>> ParseQuery(string query)
+        static IEnumerable<KeyValuePair<string, string?>> ParseQuery(string query)
             => query.Split(new[] { '&' }, StringSplitOptions.RemoveEmptyEntries)
                 .Select(
                     x => {
                         var position = x.IndexOf('=');
 
                         return position > 0
-                            ? new KeyValuePair<string, string>(x.Substring(0, position), x.Substring(position + 1))
-                            : new KeyValuePair<string, string>(x, string.Empty);
+                            ? new KeyValuePair<string, string?>(x.Substring(0, position), x.Substring(position + 1))
+                            : new KeyValuePair<string, string?>(x, null);
                     }
                 );
     }
@@ -78,14 +80,14 @@ public class RestRequest {
     /// Always send a multipart/form-data request - even when no Files are present.
     /// </summary>
     public bool AlwaysMultipartFormData { get; set; }
-    
+
     /// <summary>
     /// When set to true, parameters in a multipart form data requests will be enclosed in
     /// quotation marks. Default is false. Enable it if the remote endpoint requires parameters
     /// to be in quotes (for example, FreshDesk API). 
     /// </summary>
     public bool MultipartFormQuoteParameters { get; set; }
-    
+
     public string? FormBoundary { get; set; }
 
     /// <summary>
@@ -98,6 +100,11 @@ public class RestRequest {
     /// Optional cookie container to use for the request. If not set, cookies are not passed.
     /// </summary>
     public CookieContainer? CookieContainer { get; set; }
+
+    /// <summary>
+    /// Request-level authenticator. It will be used if set, otherwise RestClient.Authenticator will be used.
+    /// </summary>
+    public IAuthenticator? Authenticator { get; set; }
 
     /// <summary>
     /// Container of all the files to be uploaded with the request.
@@ -188,11 +195,10 @@ public class RestRequest {
     /// <summary>
     /// Set this to handle the response stream yourself, based on the response details
     /// </summary>
-    public Func<HttpResponseMessage, RestResponse>? AdvancedResponseWriter {
+    public Func<HttpResponseMessage, RestRequest, RestResponse>? AdvancedResponseWriter {
         get => _advancedResponseHandler;
         init {
-            if (ResponseWriter != null)
-                throw new ArgumentException("ResponseWriter is not null. Only one response writer can be used.");
+            if (ResponseWriter != null) throw new ArgumentException("ResponseWriter is not null. Only one response writer can be used.");
 
             _advancedResponseHandler = value;
         }
