@@ -4,8 +4,8 @@ namespace RestSharp.Serializers.CsvHelper;
 
 [PublicAPI]
 public static class RestClientExtensions {
-    public static RestClient UseCsvHelper(this RestClient client) => client.UseSerializer<CsvHelperSerializer>();
+    public static SerializerConfig UseCsvHelper(this SerializerConfig config) => config.UseSerializer<CsvHelperSerializer>();
 
-    public static RestClient UseCsvHelper(this RestClient client, CsvConfiguration configuration)
-        => client.UseSerializer(() => new CsvHelperSerializer(configuration));
+    public static SerializerConfig UseCsvHelper(this SerializerConfig config, CsvConfiguration configuration)
+        => config.UseSerializer(() => new CsvHelperSerializer(configuration));
 }
