@@ -15,13 +15,12 @@
 namespace RestSharp;
 
 static class Ensure {
-    public static T NotNull<T>(T? value, string name)
-        => value ?? throw new ArgumentNullException(name);
+    public static T NotNull<T>(T? value, [InvokerParameterName] string name) => value ?? throw new ArgumentNullException(name);
 
-    public static string NotEmpty(string? value, string name)
+    public static string NotEmpty(string? value, [InvokerParameterName] string name)
         => string.IsNullOrWhiteSpace(value) ? throw new ArgumentNullException(name) : value!;
 
-    public static string NotEmptyString(object? value, string name) {
+    public static string NotEmptyString(object? value, [InvokerParameterName] string name) {
         var s = value as string ?? value?.ToString();
         return string.IsNullOrWhiteSpace(s) ? throw new ArgumentNullException(name) : s!;
     }
