@@ -17,7 +17,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Web;
 
-namespace RestSharp.Extensions; 
+namespace RestSharp.Extensions;
 
 static class StringExtensions {
     static readonly Regex IsUpperCaseRegex = new(@"^[A-Z]+$");
@@ -33,20 +33,19 @@ static class StringExtensions {
     static readonly Regex AddSpacesRegex1 = new(@"[-\s]");
     static readonly Regex AddSpacesRegex2 = new(@"([a-z\d])([A-Z])");
     static readonly Regex AddSpacesRegex3 = new(@"([A-Z]+)([A-Z][a-z])");
-    public static string UrlDecode(this string input) => HttpUtility.UrlDecode(input);
+
+    internal static string UrlDecode(this string input) => HttpUtility.UrlDecode(input);
 
     /// <summary>
     /// Uses Uri.EscapeDataString() based on recommendations on MSDN
     /// http://blogs.msdn.com/b/yangxind/archive/2006/11/09/don-t-use-net-system-uri-unescapedatastring-in-url-decoding.aspx
     /// </summary>
-    public static string UrlEncode(this string input) {
+    internal static string UrlEncode(this string input) {
         const int maxLength = 32766;
 
-        if (input == null)
-            throw new ArgumentNullException(nameof(input));
+        if (input == null) throw new ArgumentNullException(nameof(input));
 
-        if (input.Length <= maxLength)
-            return Uri.EscapeDataString(input);
+        if (input.Length <= maxLength) return Uri.EscapeDataString(input);
 
         var sb    = new StringBuilder(input.Length * 2);
         var index = 0;
@@ -67,37 +66,18 @@ static class StringExtensions {
         return sb.ToString();
     }
 
-    public static string? UrlEncode(this string input, Encoding encoding) {
+    internal static string? UrlEncode(this string input, Encoding encoding) {
         var encoded = HttpUtility.UrlEncode(input, encoding);
         return encoded?.Replace("+", "%20");
     }
 
-    /// <summary>
-    /// Remove underscores from a string
-    /// </summary>
-    /// <param name="input">String to process</param>
-    /// <returns>string</returns>
-    public static string RemoveUnderscoresAndDashes(this string input) => input.Replace("_", "").Replace("-", "");
+    internal static string RemoveUnderscoresAndDashes(this string input) => input.Replace("_", "").Replace("-", "");
 
-    /// <summary>
-    /// Converts a string to pascal case
-    /// </summary>
-    /// <param name="lowercaseAndUnderscoredWord">String to convert</param>
-    /// <param name="culture"></param>
-    /// <returns>string</returns>
-    public static string ToPascalCase(this string lowercaseAndUnderscoredWord, CultureInfo culture)
+    internal static string ToPascalCase(this string lowercaseAndUnderscoredWord, CultureInfo culture)
         => ToPascalCase(lowercaseAndUnderscoredWord, true, culture);
 
-    /// <summary>
-    /// Converts a string to pascal case with the option to remove underscores
-    /// </summary>
-    /// <param name="text">String to convert</param>
-    /// <param name="removeUnderscores">Option to remove underscores</param>
-    /// <param name="culture"></param>
-    /// <returns></returns>
-    public static string ToPascalCase(this string text, bool removeUnderscores, CultureInfo culture) {
-        if (string.IsNullOrEmpty(text))
-            return text;
+    internal static string ToPascalCase(this string text, bool removeUnderscores, CultureInfo culture) {
+        if (string.IsNullOrEmpty(text)) return text;
 
         text = text.Replace('_', ' ');
 
@@ -113,69 +93,17 @@ static class StringExtensions {
             var restOfWord = word.Substring(1);
             var firstChar  = char.ToUpper(word[0], culture);
 
-            if (restOfWord.IsUpperCase())
-                restOfWord = restOfWord.ToLower(culture);
+            if (restOfWord.IsUpperCase()) restOfWord = restOfWord.ToLower(culture);
 
             return string.Concat(firstChar, restOfWord);
         }
     }
 
-    /// <summary>
-    /// Converts a string to camel case
-    /// </summary>
-    /// <param name="lowercaseAndUnderscoredWord">String to convert</param>
-    /// <param name="culture"></param>
-    /// <returns>String</returns>
-    public static string ToCamelCase(this string lowercaseAndUnderscoredWord, CultureInfo culture)
+    internal static string ToCamelCase(this string lowercaseAndUnderscoredWord, CultureInfo culture)
         => MakeInitialLowerCase(ToPascalCase(lowercaseAndUnderscoredWord, culture), culture);
 
-    static string MakeInitialLowerCase(this string word, CultureInfo culture)
-        => string.Concat(word.Substring(0, 1).ToLower(culture), word.Substring(1));
-
-    static string AddUnderscores(this string pascalCasedWord)
-        => AddUnderscoresRegex1.Replace(
-            AddUnderscoresRegex2.Replace(
-                AddUnderscoresRegex3.Replace(pascalCasedWord, "$1_$2"),
-                "$1_$2"
-            ),
-            "_"
-        );
-
-    static string AddDashes(this string pascalCasedWord)
-        => AddDashesRegex1.Replace(
-            AddDashesRegex2.Replace(
-                AddDashesRegex3.Replace(pascalCasedWord, "$1-$2"),
-                "$1-$2"
-            ),
-            "-"
-        );
-
-    static bool IsUpperCase(this string inputString) => IsUpperCaseRegex.IsMatch(inputString);
-
-    static string AddUnderscorePrefix(this string pascalCasedWord) => $"_{pascalCasedWord}";
-
-    static string AddSpaces(this string pascalCasedWord)
-        => AddSpacesRegex1.Replace(
-            AddSpacesRegex2.Replace(
-                AddSpacesRegex3.Replace(pascalCasedWord, "$1 $2"),
-                "$1 $2"
-            ),
-            " "
-        );
-
-    internal static bool IsEmpty(this string? value) => string.IsNullOrWhiteSpace(value);
-
-    internal static bool IsNotEmpty(this string? value) => !string.IsNullOrWhiteSpace(value);
-
-    /// <summary>
-    /// Return possible variants of a name for name matching.
-    /// </summary>
-    /// <param name="name">String to convert</param>
-    /// <param name="culture">The culture to use for conversion</param>
-    /// <returns>IEnumerable&lt;string&gt;</returns>
-    public static IEnumerable<string> GetNameVariants(this string name, CultureInfo culture) {
-        if (string.IsNullOrEmpty(name))
-            yield break;
+    internal static IEnumerable<string> GetNameVariants(this string name, CultureInfo culture) {
+        if (string.IsNullOrEmpty(name)) yield break;
 
         yield return name;
 
@@ -216,8 +144,46 @@ static class StringExtensions {
         yield return name.AddSpaces().ToLower(culture);
     }
 
+    internal static bool IsEmpty(this string? value) => string.IsNullOrWhiteSpace(value);
+
+    internal static bool IsNotEmpty(this string? value) => !string.IsNullOrWhiteSpace(value);
+
     internal static string JoinToString<T>(this IEnumerable<T> collection, string separator, Func<T, string> getString)
         => JoinToString(collection.Select(getString), separator);
 
     internal static string JoinToString(this IEnumerable<string> strings, string separator) => string.Join(separator, strings);
+
+    static string MakeInitialLowerCase(this string word, CultureInfo culture)
+        => string.Concat(word.Substring(0, 1).ToLower(culture), word.Substring(1));
+
+    static string AddUnderscores(this string pascalCasedWord)
+        => AddUnderscoresRegex1.Replace(
+            AddUnderscoresRegex2.Replace(
+                AddUnderscoresRegex3.Replace(pascalCasedWord, "$1_$2"),
+                "$1_$2"
+            ),
+            "_"
+        );
+
+    static string AddDashes(this string pascalCasedWord)
+        => AddDashesRegex1.Replace(
+            AddDashesRegex2.Replace(
+                AddDashesRegex3.Replace(pascalCasedWord, "$1-$2"),
+                "$1-$2"
+            ),
+            "-"
+        );
+
+    static bool IsUpperCase(this string inputString) => IsUpperCaseRegex.IsMatch(inputString);
+
+    static string AddUnderscorePrefix(this string pascalCasedWord) => $"_{pascalCasedWord}";
+
+    static string AddSpaces(this string pascalCasedWord)
+        => AddSpacesRegex1.Replace(
+            AddSpacesRegex2.Replace(
+                AddSpacesRegex3.Replace(pascalCasedWord, "$1 $2"),
+                "$1 $2"
+            ),
+            " "
+        );
 }
