@@ -21,7 +21,7 @@ public abstract record Parameter(string? Name, object? Value, ParameterType Type
     /// <summary>
     /// MIME content type of the parameter
     /// </summary>
-    public string? ContentType { get; protected init; }
+    public ContentType ContentType { get; protected init; } = ContentType.Undefined;
 
     /// <summary>
     /// Return a human-readable representation of this parameter
@@ -32,10 +32,10 @@ public abstract record Parameter(string? Name, object? Value, ParameterType Type
     public static Parameter CreateParameter(string? name, object? value, ParameterType type, bool encode = true)
         // ReSharper disable once SwitchExpressionHandlesSomeKnownEnumValuesWithExceptionInDefault
         => type switch {
-            ParameterType.GetOrPost   => new GetOrPostParameter(name!, value?.ToString(), encode),
-            ParameterType.UrlSegment  => new UrlSegmentParameter(name!, value?.ToString()!, encode),
+            ParameterType.GetOrPost   => new GetOrPostParameter(Ensure.NotEmptyString(name, nameof(name)), value?.ToString(), encode),
+            ParameterType.UrlSegment  => new UrlSegmentParameter(Ensure.NotEmptyString(name, nameof(name)), value?.ToString()!, encode),
             ParameterType.HttpHeader  => new HeaderParameter(name, value?.ToString()),
-            ParameterType.QueryString => new QueryParameter(name!, value?.ToString(), encode),
+            ParameterType.QueryString => new QueryParameter(Ensure.NotEmptyString(name, nameof(name)), value?.ToString(), encode),
             _                         => throw new ArgumentOutOfRangeException(nameof(type), type, null)
         };
 }
