@@ -1,4 +1,4 @@
-#if NETSTANDARD
+#if NETSTANDARD || NETFRAMEWORK
 using System.ComponentModel;
 
 // ReSharper disable once CheckNamespace
