@@ -15,6 +15,8 @@
 
 namespace RestSharp; 
 
+using System.Diagnostics.CodeAnalysis;
+
 static class BodyExtensions {
     public static bool TryGetBodyParameter(this RestRequest request, out BodyParameter? bodyParameter) {
         bodyParameter = request.Parameters.FirstOrDefault(p => p.Type == ParameterType.RequestBody) as BodyParameter;
@@ -23,5 +25,5 @@ static class BodyExtensions {
 
     public static bool HasFiles(this RestRequest request) => request.Files.Count > 0;
 
-    public static bool IsEmpty(this ParametersCollection? parameters) => parameters == null || parameters.Count == 0;
+    public static bool IsEmpty([NotNullWhen(false)]this ParametersCollection? parameters) => parameters == null || parameters.Count == 0;
 }
