@@ -15,7 +15,7 @@
 namespace RestSharp.Serializers;
 
 public interface ISerializer {
-    string ContentType { get; set; }
+    ContentType ContentType { get; set; }
 
     string? Serialize(object obj);
 }
