@@ -27,11 +27,10 @@ public class TwitterClient : ITwitterClient, IDisposable {
     readonly RestClient _client;
 
     public TwitterClient(string apiKey, string apiKeySecret) {
-        var options = new RestClientOptions("https://api.twitter.com/2");
-
-        _client = new RestClient(options) {
+        var options = new RestClientOptions("https://api.twitter.com/2") {
             Authenticator = new TwitterAuthenticator("https://api.twitter.com", apiKey, apiKeySecret)
         };
+        _client = new RestClient(options);
     }
 
     public async Task<TwitterUser> GetUser(string user) {
@@ -93,11 +92,11 @@ class TwitterAuthenticator : AuthenticatorBase {
     }
 
     async Task<string> GetToken() {
-        var options = new RestClientOptions(_baseUrl);
-
-        using var client = new RestClient(options) {
+        var options = new RestClientOptions(_baseUrl) {
             Authenticator = new HttpBasicAuthenticator(_clientId, _clientSecret),
         };
+
+        using var client = new RestClient(options);
 
         var request = new RestRequest("oauth2/token")
             .AddParameter("grant_type", "client_credentials");
