@@ -6,6 +6,6 @@ record TestServerResponse(string Name, string Value);
 
 record UploadRequest(string Filename, IFormFile File);
 
-record UploadResponse(string FileName, long Length, bool Equal);
+public record UploadResponse(string FileName, long Length, bool Equal);
 
 record ContentResponse(string Content);
