@@ -1,13 +1,18 @@
+using System.Net;
 using RestSharp.Tests.Integrated.Server;
 
 namespace RestSharp.Tests.Integrated;
 
 [Collection(nameof(TestServerCollection))]
 public class UploadFileTests {
-    readonly RestClient _client;
-    readonly string     _path = AppDomain.CurrentDomain.BaseDirectory;
+    readonly ITestOutputHelper _output;
+    readonly RestClient        _client;
+    readonly string            _path = AppDomain.CurrentDomain.BaseDirectory;
 
-    public UploadFileTests(TestServerFixture fixture) => _client = new RestClient(fixture.Server.Url);
+    public UploadFileTests(TestServerFixture fixture, ITestOutputHelper output) {
+        _output = output;
+        _client = new RestClient(new RestClientOptions(fixture.Server.Url) { ThrowOnAnyError = true });
+    }
 
     [Fact]
     public async Task Should_upload_from_file() {
@@ -16,11 +21,14 @@ public class UploadFileTests {
         var path = Path.Combine(_path, "Assets", filename);
 
         var request  = new RestRequest("upload").AddFile("file", path);
-        var response = await _client.PostAsync<UploadResponse>(request);
+        var response = await _client.ExecutePostAsync<UploadResponse>(request);
+
+        response.StatusCode.Should().Be(HttpStatusCode.OK);
 
         var expected = new UploadResponse(filename, new FileInfo(path).Length, true);
 
-        response.Should().BeEquivalentTo(expected);
+        _output.WriteLine(response.Content);
+        response.Data.Should().BeEquivalentTo(expected);
     }
 
     [Fact]
@@ -31,11 +39,12 @@ public class UploadFileTests {
         var bytes = await File.ReadAllBytesAsync(path);
 
         var request  = new RestRequest("upload").AddFile("file", bytes, filename);
-        var response = await _client.PostAsync<UploadResponse>(request);
+        var response = await _client.ExecutePostAsync<UploadResponse>(request);
 
         var expected = new UploadResponse(filename, new FileInfo(path).Length, true);
 
-        response.Should().BeEquivalentTo(expected);
+        _output.WriteLine(response.Content);
+        response.Data.Should().BeEquivalentTo(expected);
     }
 
     [Fact]
@@ -45,10 +54,11 @@ public class UploadFileTests {
         var path = Path.Combine(_path, "Assets", filename);
 
         var request  = new RestRequest("upload").AddFile("file", () => File.OpenRead(path), filename);
-        var response = await _client.PostAsync<UploadResponse>(request);
+        var response = await _client.ExecutePostAsync<UploadResponse>(request);
 
         var expected = new UploadResponse(filename, new FileInfo(path).Length, true);
 
-        response.Should().BeEquivalentTo(expected);
+        _output.WriteLine(response.Content);
+        response.Data.Should().BeEquivalentTo(expected);
     }
 }
