@@ -6,7 +6,7 @@ using RestSharp.Serializers.NewtonsoftJson;
 using RestSharp.Tests.Shared.Extensions;
 using RestSharp.Tests.Shared.Fixtures;
 
-namespace RestSharp.Tests.Serializers.Json; 
+namespace RestSharp.Tests.Serializers.Json;
 
 public class NewtonsoftJsonTests {
     static readonly Fixture Fixture = new();
@@ -66,12 +66,12 @@ public class NewtonsoftJsonTests {
 
         var testData = Fixture.Create<TestClass>();
 
-        var client  = new RestClient(server.Url).UseNewtonsoftJson();
+        var client  = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
         var request = new RestRequest().AddJsonBody(testData);
 
         await client.PostAsync(request);
 
-        var actual = serializer.Deserialize<TestClass>(new RestResponse { Content = _body! });
+        var actual = serializer.Deserialize<TestClass>(new RestResponse(request) { Content = _body! });
 
         actual.Should().BeEquivalentTo(testData);
     }
@@ -90,7 +90,7 @@ public class NewtonsoftJsonTests {
             }
         );
 
-        var client = new RestClient(server.Url).UseNewtonsoftJson();
+        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
 
         var actual = await client.GetAsync<TestClass>(new RestRequest());
 
@@ -98,18 +98,17 @@ public class NewtonsoftJsonTests {
     }
 
     [Fact]
-    public async Task DeserilizationFails_IsSuccessfull_Should_BeFalse() 
-    {
+    public async Task DeserilizationFails_IsSuccessful_Should_BeFalse() {
         using var server = HttpServerFixture.StartServer(
             (_, response) => {
-                response.StatusCode = (int)HttpStatusCode.OK;
-                response.ContentType = "application/json";
+                response.StatusCode      = (int)HttpStatusCode.OK;
+                response.ContentType     = "application/json";
                 response.ContentEncoding = Encoding.UTF8;
                 response.OutputStream.WriteStringUtf8("invalid json");
             }
         );
 
-        var client = new RestClient(server.Url).UseNewtonsoftJson();
+        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
 
         var response = await client.ExecuteAsync<TestClass>(new RestRequest());
 
@@ -118,21 +117,21 @@ public class NewtonsoftJsonTests {
     }
 
     [Fact]
-    public async Task DeserilizationSucceeds_IsSuccessfull_Should_BeTrue() {
+    public async Task DeserilizationSucceeds_IsSuccessful_Should_BeTrue() {
         var item = Fixture.Create<TestClass>();
 
         using var server = HttpServerFixture.StartServer(
             (_, response) => {
                 var serializer = new JsonNetSerializer();
 
-                response.StatusCode = (int)HttpStatusCode.OK;
-                response.ContentType = "application/json";
+                response.StatusCode      = (int)HttpStatusCode.OK;
+                response.ContentType     = "application/json";
                 response.ContentEncoding = Encoding.UTF8;
                 response.OutputStream.WriteStringUtf8(serializer.Serialize(item)!);
             }
         );
 
-        var client = new RestClient(server.Url).UseNewtonsoftJson();
+        var client = new RestClient(server.Url, configureSerialization: cfg => cfg.UseNewtonsoftJson());
 
         var response = await client.ExecuteAsync<TestClass>(new RestRequest());
 
