@@ -1,5 +1,3 @@
-using System.Globalization;
-
 namespace RestSharp.Tests;
 
 public class ObjectParserTests {
@@ -31,7 +29,7 @@ public class ObjectParserTests {
             SomeIds = new[] { 1, 2, 3 }
         };
         var expected = request.SomeIds.Select(x => ("ids[]", x.ToString()));
-        var parsed   = request.GetProperties();
+        var parsed   = request.GetProperties().Select(x => (x.Name, x.Value));
 
         parsed.Should().BeEquivalentTo(expected);
     }
